azure_data_explorer_list_databases/requirements.txt
azure-kusto-data>=4.5.0
azure-identity>=1.15.0
